Pull request #2: support for pre-trained models; readme; bugfix
Merge in CSID/glaucus

Squashed commit of the following:

commit df0d39286755416f3e570e0646d1720e47037125
Author: Kyle A Logue <kyle.a.logue@aero.org>
Date:   Thu Mar 2 17:06:05 2023 -0800

    support for pre-trained models; readme; bugfix

+++ glaucus/gblocks.py
@@ -185,7 +185,7 @@ class GBlock(pl.LightningModule):
         self.bn_mom = 1e-2 # better than torch default
         self.bn_eps = 1e-3 # better than torch default
         self.filters_ex = self.filters_in * self.expand_ratio
-        self.filters_sq = max(1, round(self.filters_ex / self.squeeze_ratio))
+        self.filters_sq = int(max(1, round(self.filters_ex / self.squeeze_ratio)))
         self.is_reshaped = self.filters_in != self.filters_out or self.stride != 1
         self.kernel_size = kernel_size
         if kernel_size % 2 == 1:
